+++ src/main/java/kr/carz/savecar/dto/YearlyRentDTO.java
@@ -16,9 +16,9 @@ public class YearlyRentDTO {
     private String category2; // 중형차, 경차 등 큰 분류
     private String deposit;
     private String name;
-    private Float cost_for_20k;
-    private Float cost_for_30k;
-    private Float cost_for_40k;
+    private double cost_for_20k;
+    private double cost_for_30k;
+    private double cost_for_40k;
     private String cost_for_others;
     private String age_limit;
     private String cost_per_km;
